GitGudWeb/classify/requirements.txt
torch
torchvision
numpy
opencv-python
pillow
ultralytics
transformers
tqdm